Run and Debug View on VS Code
In the first code we had "if(x = 10) { console.log('x is equal to 10)}, no in the terminal we ran 'node example.js' and we've got the output 'x is equal to 10' eventhough we said earlyer let x = 6. So to debug in VSCode, we put a breakpoint in th line were we declared the variable 'x', than we selected the icon to debug and clicked on 'Run and Debug' we selected the option 'Node' and we clicked to step to the next line and we could see x was equal to 6, so we clicked again to step into the next line and we could see x was equal to 10, so analizing the previous line we could realized that the syntax 'if (x = 10)' was wrong as using ''let' to declared the variable 'x', we could reassign the variable, and that is what we did. To fix, whe changed the condition as so 'if (x == 10)'.