Error found
In the console we saw a 404, in the Network tab we saw its type fetch whitch means the endpoint is not available. Upon clicking the Name of the endpoint we inpspect the Headers, Payload and Response confirming the endpoint we are POSTing the to the server is not correct. Back to VSCode and checking the server folder in user.js file we could see the correct endpoint and so fix the fetch request in skills.js file.